TID-31/new bit map font rendering (#92)
* Added hash table, buffer2d_view and buffer2d

* Integrated Freetype and added test for glyph bitmap packing

* Updated submodules

* Added TEXTURE_SAMPLING_ALPHA_CHANNEL and BITMAP_GLYPH_ATLAS_TEXTURE tests

* Fixed texture recreation upon buffer2d resize

* Implemented vulkan_bitmap_text_t, although the text is not correctly displayed

* Added SHADER_LIBRARY_SHADER_PRESET_BITMAP_TEXT
However, the bitmap text rendering is hard coded for now.
We would need to add more support the V3D shader language to make bitmap font rendering customizable.
Also, we're currently not using the screen projection matrix so the bitmap text might not scale correctly.

+++ include/renderer/internal/vulkan/vulkan_vertex_buffer_layout_description.h
@@ -335,7 +335,7 @@ typedef struct vulkan_vertex_buffer_layout_description_t
 } vulkan_vertex_buffer_layout_description_t;
 
 RENDERER_API void vulkan_vertex_buffer_layout_description_begin(vulkan_renderer_t* renderer, vulkan_vertex_buffer_layout_description_t* description, u32 stride, VkVertexInputRate input_rate, u32 binding_number);
-RENDERER_API void vulkan_vertex_buffer_layout_description_add_attribute(vulkan_vertex_buffer_layout_description_t* description, u16 location, VkFormat format, u32 offset);
+RENDERER_API void vulkan_vertex_buffer_layout_description_add_attribute(vulkan_vertex_buffer_layout_description_t* description, u32 location, VkFormat format, u32 offset);
 RENDERER_API void vulkan_vertex_buffer_layout_description_end(vulkan_renderer_t* renderer, vulkan_vertex_buffer_layout_description_t* description);
 
 static INLINE_IF_RELEASE_MODE void begin_vertex_binding(vulkan_renderer_t* renderer, BUFFER* list, u32 stride, VkVertexInputRate input_rate, u32 binding_number)
